+++ seanime-web/src/app/(main)/(library)/_containers/library-toolbar.tsx
@@ -7,12 +7,12 @@ import { __playlists_modalOpenAtom } from "@/app/(main)/(library)/_containers/pl
 import { __scanner_modalIsOpen } from "@/app/(main)/(library)/_containers/scanner-modal"
 import { __unknownMedia_drawerIsOpen } from "@/app/(main)/(library)/_containers/unknown-media-manager"
 import { __unmatchedFileManagerIsOpen } from "@/app/(main)/(library)/_containers/unmatched-file-manager"
-import { serverStatusAtom } from "@/app/(main)/_atoms/server-status.atoms"
+import { useServerStatus } from "@/app/(main)/_hooks/use-server-status"
 import { Button, IconButton } from "@/components/ui/button"
 import { cn } from "@/components/ui/core/styling"
 import { DropdownMenu, DropdownMenuItem } from "@/components/ui/dropdown-menu"
 import { Tooltip } from "@/components/ui/tooltip"
-import { useAtomValue, useSetAtom } from "jotai/react"
+import { useSetAtom } from "jotai/react"
 import Link from "next/link"
 import React from "react"
 import { BiCollection, BiDotsVerticalRounded, BiFolder } from "react-icons/bi"
@@ -34,7 +34,7 @@ export function LibraryToolbar(props: LibraryToolbarProps) {
 
     const setBulkActionIsOpen = useSetAtom(__bulkAction_modalAtomIsOpen)
 
-    const status = useAtomValue(serverStatusAtom)
+    const status = useServerStatus()
     const setScannerModalOpen = useSetAtom(__scanner_modalIsOpen)
     const setUnmatchedFileManagerOpen = useSetAtom(__unmatchedFileManagerIsOpen)
     const setUnknownMediaManagerOpen = useSetAtom(__unknownMedia_drawerIsOpen)